Copying the _print_server method from novaclient to our extension
Fixes OS-76.

The _print_server method had a signature change and now our extension
does not work with newer version of novaclient. Since this is an internal
method we should not rely on it being stable. To combat this we simply
pulled in the functionality so that we can guard against changes in the
future.

Change-Id: I52222f82dfab030a8aac81c4cc6daa908bbf4e32